db files in the indexedDB folder
